Added availability_topic and device_class
Added availability_topic to each disc_payload to inform Home Assistant of the device’s availability.

The code now checks for specific keywords in the parameter name (like "temperature", "voltage", etc.) to assign a device_class. This categorizes the entity in Home Assistant, helping it recognize the type of sensor